+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fix inheriting sampling decision from parent (#1100)
 * Fixes and Tests: Session serialization and deserialization
+* Fix: Exception only sets a stack trace if there are frames
 
 # 4.0.0-alpha.2